Docker: No longer read `env` file during build and `VERSION` is optional
- Instead of adding the `env` file in the built image, it should be
  provided during a run. Alternatively, use parameters for Docker run or
  Compose environment sections. This avoids adding organization-specific
  data (aside from the Jira/VSTS field specifications) into the image as
  well as attempting to copy a virtual environment with that name during
  the build.
- Include the vsts_fields.json file in the image build
- Allow `VERSION` file to be missing during build, which then degrades
  to the module version. This means the agent cannot report its exact
  branch/commit if the file was not added during the build.
- Document better how `env` file configuration works, how to write the
  `VERSION` file and what means there are to start the instance with
  environment variables, including Docker Compose examples elsewhere.

+++ scraper/agent/start.sh
@@ -19,7 +19,11 @@
 # limitations under the License.
 
 # Make the agent version available on the shared config volume.
-cp /home/agent/VERSION /home/agent/config/VERSION
+if [ -f /home/agent/VERSION ]; then
+	cp /home/agent/VERSION /home/agent/config/VERSION
+else
+	python -c "from gatherer import __version__;print(__version__)" > /home/agent/config/VERSION
+fi
 
 # Update configuration based on docker compose environment variables.
 (find /home/agent -name '*.cfg.example' | while read file; do
